Relocate Reigster.tsx
