+++ invitation-frontend/src/App.tsx
@@ -3,7 +3,7 @@ import './style/style.css';
 import CardBuilder from './CardBuilder';
 import Landing from './shoppingMall/Landing';
 import Login from './component/Login';
-import Register from './component/Register';
+import Register from './shoppingMall/Register';
 import HanSung from './component/HanSung';
 import HanSungList from './component/HanSungList';
 import * as Constants from './utils/Constants';
